requirements.txt
fastapi
uvicorn
pandas
numpy
torch
torchvision
pillow
requests
python-multipart

